Reversed changes away from st.navigation, changes to file names and README.md
