+++ README.md
@@ -1,8 +1,8 @@
-# Recommended reading
+# Recommended Reading
 Best practice guide: https://karpathy.github.io/2019/04/25/recipe/
 How YOLO training dataset is structured: https://docs.ultralytics.com/datasets/segment/coco8-seg/
 
-# Fine tune model
+# Fine-Tuned Models
 ## YOLO
 https://github.com/airacingtech/YOLOv8-Fine-Tune
 https://docs.ultralytics.com/datasets/segment/crack-seg/
@@ -15,6 +15,7 @@ https://github.com/biankatpas/Cracks-and-Potholes-in-Road-Images-Dataset
 https://docs.ultralytics.com/datasets/segment/crack-seg/
 https://universe.roboflow.com/search?q=model%253Ayolov11%2520potholes
 
-
-# How to run app
-streamlit run app.py 
+# How to Run Locally
+git clone https://github.com/klurpicolo/potholes-segmentation.git
+cd potholes-segmentation
+streamlit run 1_🔒_Login.py
